requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# Machine Learning
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.3.0

# LLM API
anthropic>=0.18.0

# Scientific computing
scipy>=1.11.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0

# Development
jupyter>=1.0.0
ipykernel>=6.25.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
